+++ src/main/java/com/virtustream/restful/fibonacci/FibonacciSequence.java
@@ -1,9 +1,10 @@
 package com.virtustream.restful.fibonacci;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface FibonacciSequence {
 
-    List<Integer> generate(int series);
+    List<BigInteger> generate(int series);
 
 }
